Index de acciones del front
Agrego variables de entorno al index del front, por favor agregar la línea REACT_APP_END_POINT="https://mrsemsqfk6.execute-api.us-east-1.amazonaws.com"
al archivo local .env
